fix: stop Vfs workers

+++ src/libcommonserver/vfs/vfs.cpp
@@ -37,7 +37,7 @@ Vfs::Vfs(const VfsSetupParams &vfsSetupParams, QObject *parent) :
     QObject(parent),
     _vfsSetupParams(vfsSetupParams) {}
 
-void Vfs::starVfsWorkers() {
+void Vfs::startVfsWorkers() {
     // Start hydration/dehydration workers
     // !!! Disabled for testing because no QEventLoop !!!
     if (qApp) {
@@ -57,35 +57,36 @@ void Vfs::starVfsWorkers() {
     }
 }
 
-Vfs::~Vfs() {
+void Vfs::stopVfsWorkers() {
+    LOG_DEBUG(logger(), "Stop VFS workers for syncdbid=" << _vfsSetupParams.syncDbId);
+
     // Ask workers to stop
     for (auto &worker: _workerInfo) {
         worker._mutex.lock();
-        worker._stop = true;
-        worker._mutex.unlock();
-        worker._queueWC.wakeAll();
+        if (!worker._stop) {
+            worker._stop = true;
+            worker._mutex.unlock();
+            worker._queueWC.wakeAll();
+        }
     }
 
-    // Ask workers' threads to quit
+    Utility::msleep(100);
+
     for (auto &worker: _workerInfo) {
+        // Force threads to stop if needed
         for (QThread *const workerThread: std::as_const(worker._threadList)) {
-            if (workerThread) {
-                workerThread->quit();
+            if (workerThread && workerThread->isRunning()) {
+                workerThread->terminate();
+                workerThread->wait();
             }
         }
     }
 
-    // Terminate workers' threads
-    for (auto &worker: _workerInfo) {
-        for (QThread *const workerThread: std::as_const(worker._threadList)) {
-            if (workerThread) {
-                if (!workerThread->wait(1000)) {
-                    workerThread->terminate();
-                    workerThread->wait();
-                }
-            }
-        }
-    }
+    LOG_DEBUG(logger(), "Stop VFS workers for syncdbid=" << _vfsSetupParams.syncDbId << " done");
+}
+
+Vfs::~Vfs() {
+    LOG_DEBUG(logger(), "~Vfs");
 }
 
 CommString Vfs::modeToString(KDC::VirtualFileMode virtualFileMode) {
@@ -128,6 +129,8 @@ ExitInfo Vfs::start(bool &installationDone, bool &activationDone, bool &connecti
 }
 
 void Vfs::stop(bool unregister) {
+    stopVfsWorkers();
+
     if (_started) {
         stopImpl(unregister);
         _started = false;
